Improve code
These are a bunch of changes. None is clearly necessary, it's just
proposals or inspirations.

 * Consistent use of const/let (let is fine but it seemed random)
 * Consistent way of finding DOM elements and handling errors
 * Using automatic return types of functions
 * Remove unnecessary try-catch-throw with no handling
 * Add types for API response

Typing the API response already is a good practice as it provides type-
safety as early as possible. I used a library I found because the payload
structure is quite complex. Found a mistake in the previous code as well:
`temp` is a number, not a string.

As you can see, TypeScript is quite clever in inferring types from code,
such as the return values of functions. It makes it less code to type
and update.